fix(hw14): resolve discuss. fix DI. change configuration
