+++ hw12_13_14_15_calendar/cmd/scheduler/config.go
@@ -5,15 +5,15 @@ import (
 
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/config"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/logger"
-	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/producer"
+	internalrmqproducer "github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/producer/rmq"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/storage"
 )
 
 type Config struct {
-	Logger   *logger.Config   `mapstructure:"logger"`
-	Database *storage.Config  `mapstructure:"database"`
-	Producer *producer.Config `mapstructure:"mq"`
-	Duration time.Duration    `mapstructure:"duration"`
+	Logger   *logger.Config              `mapstructure:"logger"`
+	Database *storage.Config             `mapstructure:"database"`
+	Producer *internalrmqproducer.Config `mapstructure:"rmq"`
+	Duration time.Duration               `mapstructure:"duration"`
 }
 
 func NewConfig(configFile string) (Config, error) {
