+++ hw12_13_14_15_calendar/cmd/sender/config.go
@@ -3,12 +3,12 @@ package main
 import (
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/config"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/logger"
-	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/consumer"
+	internalrmqconsumer "github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/consumer/rmq"
 )
 
 type Config struct {
-	Logger   *logger.Config   `mapstructure:"logger"`
-	Consumer *consumer.Config `mapstructure:"mq"`
+	Logger   *logger.Config              `mapstructure:"logger"`
+	Consumer *internalrmqconsumer.Config `mapstructure:"rmq"`
 }
 
 func NewConfig(configFile string) (Config, error) {
