+++ hw12_13_14_15_calendar/configs/config_scheduler.toml
@@ -15,7 +15,7 @@ user = "calendar"
 password = "developer"
 ssl = "disable"
 
-[mq]
+[rmq]
 host = "localhost"
 port = "5672"
 protocol = "amqp"
