+++ hw12_13_14_15_calendar/configs/config_sender.toml
@@ -3,7 +3,7 @@ level = "DEBUG"
 out_paths = ["stdout"]
 err_paths = ["stderr"]
 
-[mq]
+[rmq]
 host = "localhost"
 port = "5672"
 protocol = "amqp"
