+++ hw12_13_14_15_calendar/internal/mq/consumer/rmq/consumer.go
@@ -5,10 +5,19 @@ import (
 	"log"
 
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/logger"
-	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/consumer"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/utils"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/pkg/rmq"
 )
+
+type Config struct {
+	Host         string `mapstructure:"host"`
+	Port         string `mapstructure:"port"`
+	Protocol     string `mapstructure:"protocol"`
+	Username     string `mapstructure:"username"`
+	Password     string `mapstructure:"password"`
+	Subscription string `mapstructure:"subscription"`
+	ConsumerName string `mapstructure:"consumer_name"`
+}
 
 type Consumer struct {
 	Addr         string
@@ -18,7 +27,7 @@ type Consumer struct {
 	logger       logger.Logger
 }
 
-func New(conf *consumer.Config, logger logger.Logger) *Consumer {
+func New(conf *Config, logger logger.Logger) *Consumer {
 	addr, err := utils.GetMqAddress(conf.Protocol, conf.Host, conf.Port, conf.Username, conf.Password)
 	if err != nil {
 		log.Fatal(err)
