+++ hw12_13_14_15_calendar/internal/mq/producer/producer.go
@@ -4,16 +4,6 @@ import (
 	"context"
 )
 
-type Config struct {
-	Host         string `mapstructure:"host"`
-	Port         string `mapstructure:"port"`
-	Protocol     string `mapstructure:"protocol"`
-	Username     string `mapstructure:"username"`
-	Password     string `mapstructure:"password"`
-	QueueName    string `mapstructure:"queue"`
-	ExchangeName string `mapstructure:"exchange"`
-}
-
 type Producer interface {
 	Connect(context.Context) error
 	Close(context.Context) error
