--- hw12_13_14_15_calendar/internal/mq/producer/usecase.go
@@ -1,19 +1,0 @@
-package producer
-
-import (
-	"context"
-)
-
-type ProducerUseCase struct { //nolint:revive
-	producer Producer
-}
-
-func (p *ProducerUseCase) Publish(ctx context.Context, data []byte) error {
-	return p.producer.Publish(ctx, data)
-}
-
-func New(producer Producer) *ProducerUseCase {
-	return &ProducerUseCase{
-		producer: producer,
-	}
-}
