+++ hw12_13_14_15_calendar/internal/sender/sender.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/logger"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/consumer"
-	internalrmqconsumer "github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/mq/consumer/rmq"
 	"github.com/BBeRsErKeRR/otus-golang-hw/hw12_13_14_15_calendar/internal/storage"
 	"github.com/goccy/go-json"
 	"go.uber.org/zap"
@@ -15,13 +14,13 @@ import (
 
 type App struct {
 	logger logger.Logger
-	u      consumer.ConsumerUseCase
+	c      consumer.Consumer
 }
 
-func New(logger logger.Logger, u consumer.ConsumerUseCase) *App {
+func New(logger logger.Logger, c consumer.Consumer) *App {
 	return &App{
 		logger: logger,
-		u:      u,
+		c:      c,
 	}
 }
 
@@ -41,13 +40,5 @@ func (a *App) Consume(ctx context.Context) error {
 		)
 		a.logger.Info(notification)
 	}
-	return a.u.Consume(ctx, f)
-}
-
-func getConsumer(cfg *consumer.Config, logger logger.Logger) consumer.Consumer {
-	return internalrmqconsumer.New(cfg, logger)
-}
-
-func GetConsumerUseCase(cfg *consumer.Config, logger logger.Logger) consumer.ConsumerUseCase {
-	return *consumer.New(getConsumer(cfg, logger))
+	return a.c.Consume(ctx, f)
 }
